move kitchen_sink test files to their own directory
