+++ json_serializable/test/kitchen_sink_test_files/kitchen_sink_interface.dart
