+++ json_serializable/test/kitchen_sink_test_files/simple_object.dart
